+++ downstreamadapter/sink/cloudstorage/writer_test.go
@@ -14,7 +14,6 @@
 package cloudstorage
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"net/url"
@@ -32,7 +31,6 @@ import (
 	"github.com/pingcap/ticdc/pkg/pdutil"
 	"github.com/pingcap/ticdc/pkg/sink/cloudstorage"
 	"github.com/pingcap/ticdc/pkg/sink/codec/common"
-	"github.com/pingcap/ticdc/pkg/sink/failpointrecord"
 	"github.com/pingcap/ticdc/pkg/util"
 	"github.com/pingcap/ticdc/utils/chann"
 	timodel "github.com/pingcap/tidb/pkg/meta/model"
@@ -131,113 +129,4 @@ func TestWriterRun(t *testing.T) {
 	cancel()
 	d.close()
 	wg.Wait()
-}
-
-func TestMutateMessageValueForFailpointRecordClassification(t *testing.T) {
-	t.Parallel()
-
-	msg := &common.Message{
-		Value: []byte(
-			`{"pkNames":["id"],"data":[{"id":"1","c2":"v1"}]}` +
-				"\r\n" +
-				`{"pkNames":["id"],"data":[{"id":"2","_tidb_origin_ts":"100"}]}`,
-		),
-	}
-	rowRecords := []failpointrecord.RowRecord{
-		{
-			CommitTs:    101,
-			PrimaryKeys: map[string]any{"id": "1"},
-		},
-		{
-			CommitTs:    102,
-			PrimaryKeys: map[string]any{"id": "2"},
-		},
-	}
-
-	mutatedRows, originTsMutatedRows := mutateMessageValueForFailpoint(msg, rowRecords)
-
-	require.Len(t, mutatedRows, 1)
-	require.Equal(t, uint64(101), mutatedRows[0].CommitTs)
-	require.Equal(t, "1", mutatedRows[0].PrimaryKeys["id"])
-
-	require.Len(t, originTsMutatedRows, 1)
-	require.Equal(t, uint64(102), originTsMutatedRows[0].CommitTs)
-	require.Equal(t, "2", originTsMutatedRows[0].PrimaryKeys["id"])
-
-	require.True(t, bytes.Contains(msg.Value, []byte(`"_tidb_origin_ts":"101"`)))
-	require.True(t, bytes.Contains(msg.Value, []byte(`"c2":null`)))
-}
-
-func TestSelectColumnToMutateSkipNilOriginTsWhenPossible(t *testing.T) {
-	t.Parallel()
-
-	row := map[string]any{
-		"id":                       "1",
-		commonEvent.OriginTsColumn: nil,
-		"c2":                       "v1",
-	}
-	pkSet := map[string]struct{}{
-		"id": {},
-	}
-
-	for i := 0; i < 20; i++ {
-		col, ok := selectColumnToMutate(row, pkSet)
-		require.True(t, ok)
-		require.Equal(t, "c2", col)
-	}
-}
-
-func TestSelectColumnToMutatePreferNonNilOriginTs(t *testing.T) {
-	t.Parallel()
-
-	row := map[string]any{
-		"id":                       "1",
-		commonEvent.OriginTsColumn: "100",
-		"c2":                       "v1",
-	}
-	pkSet := map[string]struct{}{
-		"id": {},
-	}
-
-	for i := 0; i < 20; i++ {
-		col, ok := selectColumnToMutate(row, pkSet)
-		require.True(t, ok)
-		require.Equal(t, commonEvent.OriginTsColumn, col)
-	}
-}
-
-func TestSelectColumnToMutateSkipNilNonPKColumns(t *testing.T) {
-	t.Parallel()
-
-	row := map[string]any{
-		"id": "1",
-		"c1": nil,
-		"c2": "v2",
-	}
-	pkSet := map[string]struct{}{
-		"id": {},
-	}
-
-	for i := 0; i < 20; i++ {
-		col, ok := selectColumnToMutate(row, pkSet)
-		require.True(t, ok)
-		require.Equal(t, "c2", col)
-	}
-}
-
-func TestSelectColumnToMutateNoCandidateWhenAllNonPKColumnsNil(t *testing.T) {
-	t.Parallel()
-
-	row := map[string]any{
-		"id":                       "1",
-		commonEvent.OriginTsColumn: nil,
-		"c1":                       nil,
-	}
-	pkSet := map[string]struct{}{
-		"id": {},
-	}
-
-	col, ok := selectColumnToMutate(row, pkSet)
-	require.False(t, ok)
-	require.Empty(t, col)
 }
